+++ components/ProposalCard.tsx
@@ -3,7 +3,7 @@ import { ChevronRightIcon } from '@heroicons/react/solid'
 import ProposalStateBadge from './ProposalStatusBadge'
 import Link from 'next/link'
 import { Proposal, ProposalState } from '@solana/spl-governance'
-import ApprovalQuorum, { ApprovalProgress } from './ApprovalQuorum'
+import { ApprovalProgress } from './QuorumProgress'
 import useRealm from '../hooks/useRealm'
 import useProposalVotes from '../hooks/useProposalVotes'
 import ProposalTimeStatus from './ProposalTimeStatus'
