+++ components/ProposalSelectCard.tsx
@@ -1,7 +1,7 @@
 import { CheckIcon } from '@heroicons/react/solid'
 import ProposalStateBadge from './ProposalStatusBadge'
 import { Proposal, ProposalState } from '@solana/spl-governance'
-import ApprovalQuorum, { ApprovalProgress } from './ApprovalQuorum'
+import { ApprovalProgress } from './QuorumProgress'
 import useProposalVotes from '../hooks/useProposalVotes'
 import ProposalTimeStatus from './ProposalTimeStatus'
 import { PublicKey } from '@solana/web3.js'
