+++ pages/dao/[symbol]/proposal/[pk]/index.tsx
@@ -6,7 +6,7 @@ import ProposalStateBadge from 'components/ProposalStatusBadge'
 import { InstructionPanel } from 'components/instructions/instructionPanel'
 import DiscussionPanel from 'components/chat/DiscussionPanel'
 import VotePanel from 'components/VotePanel'
-import ApprovalQuorum, { ApprovalProgress } from 'components/ApprovalQuorum'
+import { ApprovalProgress } from '@components/QuorumProgress'
 import useRealm from 'hooks/useRealm'
 import useProposalVotes from 'hooks/useProposalVotes'
 import ProposalTimeStatus from 'components/ProposalTimeStatus'
